Speed up by a factor of 110×
Summary:
The problem, of course, was that I didn't read the docs closely enough
and was using a naive closure algorithm instead of Schreier–Sims!
This solves everything: time is down from 11 minutes to 6 seconds
and the result is the same.

Test Plan:
`cabal run` should print `5040` and then `Right 1044`, as usual.